% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\name{basicTheme}
\alias{basicTheme}
\title{NicePlots Theme: Basic}
\usage{
basicTheme
}
\description{
This is the default theme for nicePlots when data point coloring is active
}
\details{
This default theme has uses transparent solid circles for point overlays with up to 8 colors.
Fill and line options as constant.
}
