% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_utility.R
\name{boot95ci}
\alias{boot95ci}
\title{Calculate a basic bootstrap 95\% confidence interval}
\usage{
boot95ci(x, agg = "mean", upper = FALSE, nBoot = 1000L)
}
\arguments{
\item{x}{numeric vector}

\item{agg}{character: A string corresponding to the aggregate or central tendency function to be modeled (eg. \code{\link[stats]{median}}, \code{\link[base]{mean}}, etc.)}

\item{upper}{logical: Should the upper or lower confidence interval boundary be returned}

\item{nBoot}{numeric; An integer value indicating the number of boot strap resampling trails to run. Defaults to 1,000.}
}
\value{
a number corresponding to the upper or lower bootrap 95\% confidence interval of the aggregator function given by \code{agg}.
}
\description{
Calculates a basic 95\% boostrap confidence interval for an aggregator function
}
\details{
This function calculates a bootstrap 95\% confidence interval for a aggregator function determined by the \code{agg} variable.
The number of iterations is hard coded at 1000. The variable \code{upper} will cause to return the upper bound of the 95\%
confidence interval if set to \code{\link{TRUE}} or the lower bound if set to \code{\link{FALSE}}.
}
\examples{
library(boot)
data(iris)
median(iris$Sepal.Width)
#boot95ci(iris$Sepal.Width,agg="median",upper=FALSE)
#boot95ci(iris$Sepal.Width,agg="median",upper=TRUE)
}
\seealso{
\code{\link{prepBarData}}, \code{\link{niceBar}}, \code{\link[boot]{boot}}, \code{\link[boot]{boot.ci}}
}
