% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_utility.R
\name{calcStats}
\alias{calcStats}
\title{calculate preliminary statistical significance analysis}
\usage{
calcStats(x, by, type = c("Wilcox", "Tukey", "T.Test", "ANOVA"),
  verbose = FALSE)
}
\arguments{
\item{x}{numeric; numeric vector of data points to analyze.}

\item{by}{factor; factor describing the groups within \code{x} to test.}

\item{type}{character; determines which statistical test should be used. Accepted values are 'wilcox', 't.test', 'ttest', 'anova' and 'tukey'. Values not matching a valid input will produce a warning.}

\item{verbose}{logical; will print statistical output to the screen if set \code{\link{TRUE}}. Calculations returned by the function either way.}
}
\value{
a character string describing the test run and the p-value.
}
\description{
\code{calcStats} takes a numeric vector and a factor and runs a preliminary statistical analysis. Output is printed to the screen and the p-value is returned as a character string.
}
\details{
This is designed to be used in conjunction with data visualization plots to help with data exploration and should not be used for a robust statistical analysis. Normal distribution, variance and other data characteristics are not evaluated and there is no guarantee that the underling test assumptions are met. For two level factors \code{\link{wilcox.test}} or \code{\link{t.test}} is recommended. If the factor has more than two levels then \code{\link{pairwise.wilcox.test}} and \code{\link{pairwise.t.test}} are automatically selected. In this case \code{\link{anova}} and the optional follow-up \code{\link{TukeyHSD}} can also be used. All output it printed to the console and for the two level tests and \code{\link{anova}} the p-value is returned as a text string.
}
\examples{
data(iris)
\donttest{pv<-calcStats(iris$Petal.Length,by=iris$Species,type="anova")}
\donttest{boxplot(iris$Petal.Length~iris$Species,main="Petal Length by Species",sub=pv)}

}
\seealso{
\code{\link[stats]{wilcox.test}}, \code{\link[stats]{pairwise.wilcox.test}}, \code{\link[stats]{t.test}}, \code{\link[stats]{pairwise.t.test}}, \code{\link[stats]{anova}}, \code{\link[stats]{TukeyHSD}}
}
