% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_utility.R
\name{ci}
\alias{ci}
\title{Confidence interval helper function}
\usage{
ci(x, indices, agg)
}
\arguments{
\item{x}{numeric vector}

\item{indices}{numeric vector: Indicies are passed to calculated individual bootstaps of \code{x}}

\item{agg}{character: A string corresponding to the aggregate or central tendency function to be modeled (eg. \code{\link[stats]{median}}, \code{\link[base]{mean}}, etc.)}
}
\value{
a number corresponding to a bootstrap iteration of the aggregator function given by \code{agg}
}
\description{
Designed to be used by \code{\link{boot95ci}} to calculate a bootstrap model of an aggregator function given by \code{agg}.
}
\details{
This is a convenience function internal to \code{NicePlots} and is not exported. The variable \code{agg} should be a string corresponding
an aggregator function such as \code{\link[base]{mean}} or \code{\link[stats]{median}}. The indicies designed to passed from the \code{\link[boot]{boot}} function
from the package \code{boot} and are used to determine calculated the different bootstrap iterations of the \code{agg} functions on \code{x}.
}
\examples{
#Calculates the median of Sepal.Length from the iris data set
data(iris)
#ci(iris$Sepal.Length,"median",seq(1,length(iris$Sepal.Length)))
}
\seealso{
\code{\link[boot]{boot}}, \code{\link{boot95ci}}
}
