% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_data_processing.R
\name{dataFlightCheck}
\alias{dataFlightCheck}
\title{Check data formating for the NicePlots package}
\usage{
dataFlightCheck(data, by, flipFacts, na.rm = FALSE)
}
\arguments{
\item{data}{vector or dataframe; data to be plotted}

\item{by}{factor or dataframe; factors to be used to format data}

\item{flipFacts}{logical; If a dataframe is used for plotting data input, this will covert the data to a vector with the dataframe colums trasfered a factor in the second column of the \code{by} input.}

\item{na.rm}{logical; Removes all data and factor rows were \code{NA} is presenst.}
}
\value{
A named list with \code{d=data} and \code{b=by}.
}
\description{
Formats and cleans data prior to setting up the plotting enviroment
}
\details{
This funciton makes sure the \code{data} input is a numeric vector or a data frame of numeric vectors.
It will also check to make sure \code{by} is a factor or a dataframe of factors. If requested, it will also remove missing data and rearrage numeric dataframe inputs.
}
\examples{
todo<-1

}
