% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_plotting_functions.R
\name{drawPoints}
\alias{drawPoints}
\title{draw dots for a dot plot}
\usage{
drawPoints(
  x,
  type = "jitter",
  col = "black",
  size = 1,
  shape = 1,
  highlight = FALSE,
  width = 0.2,
  sidePlot = FALSE,
  swarmOverflow = "random",
  calcOnly = FALSE
)
}
\arguments{
\item{x}{named list or data frame; \code{x$at}, \code{x$data} and \code{x$pfact} (optional) should all be defined. These vectors are used to place the the point on the chart and determine the point level grouping (highlighting)}

\item{type}{character; determines how the points are arranged. Options are 'jitter', 'linear', 'beeswarm' and 'distribution'.}

\item{col}{character; vector of color names for plotting points. If length is greater than one it will be used for subgroups or will iterate over the groups.}

\item{size}{numeric; vector of cex values for point size. If length is greater than one it will be used for subgroups or will iterate over the groups.}

\item{shape}{numeric; vector determining point shapes (pch). If length is greater than one it will be used for subgroups or will iterate over the groups.}

\item{highlight}{logical; Should the point highlighting option be turned on (assumes that pfact is defined).}

\item{width}{numeric; determines how far points can deviate from the center category label for \code{type} options other than 'linear'.}

\item{sidePlot}{logical; plots dots for a horizontal rather than vertical axis.}

\item{swarmOverflow}{character; How to handle beeswarms that would normally overflow the \code{width} argument. Valid options are "random", "gutter", "wrap", "omit", and "none".}

\item{calcOnly}{logical; Will return just xy plotting data but will plot nothing if set to \code{TRUE}. Setting drawPoints to false will still plot outlier data.}
}
\description{
takes a data frame of locations, values and an optional subgrouping factor and adds the data points to the active plot
}
\details{
This function adds data points to a chart. These can be organized exactly as specified (linear), as a jitter cloud (jitter), as a waterfall plot (distribution) or as a swarm (beeswarm).
A factor labeled pfact can be included in \code{x} and used to highlight individual data points by setting \code{subgroup=\link{TRUE}}. All graphic customization options can given as vectors and will be iterated over during plotting.
Note that the size/cex option can not be used to highlight pfact levels in a beeswarm plot and only the first element of the vector will be used.
The function silently returns the final xy positions.
}
\examples{
data(iris)
boxplot(iris$Sepal.Length~iris$Species,ylab="Sepal Length")
iData<-data.frame(at=as.numeric(iris$Species),data=iris$Sepal.Length)
drawPoints(iData,type="jitter",col=c("red","blue","purple"))
}
\seealso{
\code{\link[graphics]{points}}, \code{\link[graphics]{stripchart}}, \code{\link[beeswarm]{beeswarm}}
}
