% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_plotting_functions.R
\name{drawViolinPlot}
\alias{drawViolinPlot}
\title{draw a violin plot}
\usage{
drawViolinPlot(x, groups, at = seq(1, length(levels(groups))),
  h = NULL, plotColors = basicTheme$plotColors, sidePlot = FALSE,
  borderCol = plotColors$lines, borderWidth = 1,
  fill = plotColors$fill, width = 1, trimViolins = TRUE,
  samplePoints = NULL)
}
\arguments{
\item{x}{numeric; A vector of numeric values that will be subset and formated by the factor(s) in \code{by}.}

\item{groups}{factor; A factor used to subset \code{x} to draw the violins.}

\item{at}{numeric; a numeric vector of where each factor level should be plotted}

\item{h}{numeric; Bandwidth for the kernal desnisty estimates. Will cycle over values if multiple bandwidths are given}

\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions.}

\item{sidePlot}{logical; If \code{\link{TRUE}}, the x and y ploting axis are swapped}

\item{borderCol}{R color string; Color of the border of the violins}

\item{borderWidth}{numeric; Thickness of the violin borders (lwd)}

\item{fill}{R color string; Color of the interior of the violins}

\item{width}{numeric; Relative width of the violins. A value of 1 will cause the violins to cover thier entire lane and potentially just touch.}

\item{trimViolins}{logical; Should the violins be truncated at the edges of the data range.}

\item{samplePoints}{integer; The number of points used to draw each side of the violin. This is generally obtained from \code{theme$curvePoints}.}
}
\description{
Produce a violin plot with optional box plot and strip plot overlays
}
\details{
This uses  \code{\link[KernSmooth]{bkde}} from the \code{KernSmooth} package to calculated kernal desnisty estimates and estimate the
optimal bandwidth \code{h} setting. This data is then used to draw a violin plot with an optional boxplot drawn as an overlay
to better charactarize the quartile distributon. Likewise, a strip chart of individual data points
can be added on top of these two plots to full charactarize the data distribution. This uses \code{\link{niceBox}} to handle
the box plot and strip chart overlays.
}
\examples{
todo<-1
}
\seealso{
\code{\link{niceVio}}, \code{\link[KernSmooth]{bkde}}, \code{\link[KernSmooth]{dpik}}
}
