% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_utility.R
\name{facetSpacing}
\alias{facetSpacing}
\title{Generate plotting locations for subgrouping data}
\usage{
facetSpacing(subGroup, labels)
}
\arguments{
\item{subGroup}{positive integer; number of levels in the subgrouping factor}

\item{labels}{positive integer; number of levels in the primary factor}
}
\value{
a numeric vector of where to plot the subgrouped data. Can be supplied to that \code{at=} option in plotting functions
}
\description{
\code{facetSpacing} generates a vector for the \code{at=} specification in functions for data sub-grouping
}
\details{
\code{facetSpacing} takes the number factor levels from the primary and secodary grouping factors to generate a vector of positions for plotting subgrouped data for the nicePlots package.
The spacing assumes that each primary factor levels is plot on positive integers 1, 2, 3 etc.
For a primary factor at position \code{i} with \code{f} subgroup levels, the subgrouping comes from generating equally spaced intervals starting at \eqn{i-\frac{1}{2}+\frac{1}{f+1}}{i-.5+1/(f+1)} and ending at \eqn{i+\frac{1}{2}-\frac{1}{f+1}}{i+.5-1/(f+1)}. Simply put: \deqn{Spacing = \frac{1}{NSubGroups-1}}
}
\examples{
\donttest{boxplot(CNA$BM~ CNA$Status,border="white")}
\donttest{stripchart(CNA$BM~factor(paste0(CNA$Status,CNA$Sex)),add=T,at=facetSpacing(2,2))}
}
\seealso{
\code{\link{prepCategoryWindow}}
}
