% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_options_processing.R
\name{formatPlotColors}
\alias{formatPlotColors}
\title{format a NicePlots color list}
\usage{
formatPlotColors(plotColors, theme = NA)
}
\arguments{
\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions. Names left unspecified will be added and set to default values automatically.}

\item{theme}{list; A \code{NicePlots} plotColor list from a theme.}
}
\value{
a formated NicePlots color list.
}
\description{
To simplify code and user options, any color option not set by the user is added to the list and set to the default value.
}
\details{
The \code{NicePlots} plotColors object is a list of named color values/vectors. The options NicePlots colors include \code{bg} (background color), \code{marginBg} (color of area surrounding the plot), \code{guides} (guide lines for major tick-marks), \code{minorGuides} (guide lines for minor tick-marks)
\code{lines} (lines for box/bar plots etc.), \code{points} (plotting data points), \code{fill} (fill for box/bar plots etc.), \code{axis} (axis colors), \code{majorTick} (major tick-mark color),
\code{minorTick} (minor tick-mark color), \code{labels} (label colors), \code{subGroupLabels} (subgroup label colors), \code{rectCol} (inner quartile range box used in \code{\link{niceVio}}), and \code{medianMarkerCol} (the median value marker used in \code{\link{niceVio}}).
Any option not set be the user will be added to the list and set to the default in order to insure compatibility with downstream NicePlot functions.
If a theme is given, any option not set by the user will be set by the theme.
}
\examples{
myCols<-list(bg="lightgrey",fill=c("red","green","blue"),lines="darkgrey")
#\donttest{myCols<-formatPlotColors(myCols)}
print(myCols)
}
