% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicePlots2.R
\name{makeColorMatrix}
\alias{makeColorMatrix}
\title{Create a matrix of increasingly transparent colors}
\usage{
makeColorMatrix()
}
\value{
A \code{6 x 5} matrix of colors.
}
\description{
\code{makeColorMatrix} is a convenience function for plotting with transparent colors.
}
\details{
This function take no arguments, but generates rows corresponding to red, blue, green, gray, purple and gold with increasing transparency moving from left to right across the columns.
}
\examples{
plot(1,1,col="white",xlim=c(0,10),ylim=c(0,10))
for(n in 1:6){rect(0:4,rep(8-n,5),1:5,rep(9-n,5),col=as.matrix(makeColorMatrix())[n,])}

#An example how it can be used in practice:
myData<-rnorm(600)
fact<-factor(c(rep("a",100),rep("b",100),rep("c",100),rep("d",100),rep("e",100),rep("f",100)))
plot(myData,col=makeColorMatrix()[fact,3])
}
\seealso{
\code{\link[grDevices]{rainbow}}, \code{\link[grDevices]{col2rgb}}, \code{\link[grDevices]{rgb}}.
}
