% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\name{newNPTheme}
\alias{newNPTheme}
\title{Creat a NicePlots Theme}
\usage{
newNPTheme(theme = basicTheme, plotColors = NULL, ...)
}
\arguments{
\item{theme}{list; A \code{NicePlots} plotColor list from a theme or a list with values corresponding to the plotColor options you wish to update.}

\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions. Names left unspecified will be added and set to default values automatically.}

\item{...}{All theme options can be set but as if they were comma separted funcitonal arguments. See example for details.}
}
\value{
A \code{NicePlots} theme of class "npTheme".
}
\description{
Makes a new \code{NicePlots} theme off of an existing template
}
\details{
Useing an exisiting template, this function updates all of the options theme options
supplied in \code{...}. To see the theme options, print an existing theme to the console.
You can see the available themes using the \code{\link{npThemes}} function.
}
\examples{
data(iris)
npThemes()
myTheme<-newNPTheme(theme=npColorTheme,
   plotColors=list(lines=c("darkgrey")),
   pointMethodBP="beeswarm")
npThemes()
niceBox(iris[,1:2],iris$Species,subgroup=TRUE,legend=TRUE,theme=myTheme)
}
\seealso{
\code{\link{npThemes}}
}
