% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_niceDensity.R
\name{niceDensity}
\alias{niceDensity}
\title{draw a kernal density plot}
\usage{
niceDensity(
  x,
  by = NULL,
  drawPoints = TRUE,
  groupNames = NULL,
  subGroup = FALSE,
  useRgl = TRUE,
  plotType = c("contour", "surface"),
  theme = basicTheme,
  main = NULL,
  sub = NULL,
  ylab = NULL,
  xlab = NULL,
  minorTick = FALSE,
  guides = NULL,
  plotColors = NULL,
  logScale = FALSE,
  axisText = c(NULL, NULL),
  showCalc = FALSE,
  calcType = "none",
  rotateLabels = FALSE,
  add = FALSE,
  minorGuides = NULL,
  extendTicks = TRUE,
  expLabels = FALSE,
  lWidth = NULL,
  na.rm = FALSE,
  verbose = FALSE,
  logAdjustment = 1,
  xLim = NULL,
  yLim = NULL,
  strictLimits = FALSE,
  legend = FALSE,
  trimCurves = TRUE,
  ...
)
}
\arguments{
\item{x}{numeric vector or data frame; The input to \code{prepCategoryWindow} can be a numeric vector a  data frame of numeric vectors.}

\item{by}{factor or data frame of factors; used as the primary grouping factor and the factor levels will be used as group names if \code{groupNames} is not specified. If \code{by} is a data frame and \code{subGroup=\link{TRUE}}, the second column is assumed to be a secondary grouping factor, breaking out the data into sub-categories within each major group determined by the levels of the first column.}

\item{drawPoints}{logical; draws a dot plot overlay for contour plots}

\item{groupNames}{character vector; overrides the factor levels of \code{by} to label the groups}

\item{subGroup}{logical; Will use the factor levels in \code{by} to plot a series of distributions from the data in \code{x}.}

\item{useRgl}{logical; Should the library \code{\link[rgl]{rgl}} be used to make 3D surface plots.}

\item{plotType}{character; Can be set to \code{contour} or \code{surface} to control the type of 2D plot.}

\item{theme}{list object; Themes are are an optional way of storing graphical preset options that are compatible with all nicePlot graphing functions.}

\item{main}{character; title for the graph which is supplied to the \code{main} argument.}

\item{sub}{character; subtitle for the graph which is supplied to the \code{sub} argument. If \code{\link{NULL}} and \code{showCalc=\link{TRUE}} it will be used to display the output form \code{\link{calcStats}}.}

\item{ylab}{character; y-axis label.}

\item{xlab}{character; x-axis label.}

\item{minorTick}{positive integer; number of minor tick-marks to draw between each pair of major ticks-marks.}

\item{guides}{logical; will draw guidelines at the major tick-marks if set to \code{\link{TRUE}}. Color of the guidelines is determined by \code{plotColors$guides}.}

\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions. Names left unspecified will be added and set to default values automatically.}

\item{logScale}{numeric; A length two numeric vector indicating the log scale the x and y axis, respectively. If only one value is given, it is applied to boxth axis. Set to \code{\link{FALSE}} to disable (Default).}

\item{axisText}{list; A named list, "x" and "y" with two optional character values to be placed before and/or after the numeric axis label, respectively.}

\item{showCalc}{logical; TODO Not sure if there are calcs worth doing here... maybe normality testing?}

\item{calcType}{charactar; TODO Not sure if there are calcs worth doing here...maybe normality testing?}

\item{rotateLabels}{logical; sets \code{las=2} for the x-axis category labels. Will affect y-axis if \code{side=\link{TRUE}}. Note that this may not work well if long names or with subgrouped data.}

\item{add}{logical; Adds the plot to the existing window rather the generating a new plotting enviroment}

\item{minorGuides}{logical; draws guidelines at minor tick-marks}

\item{extendTicks}{logical; extends minor tick-marks past the first and last major tick to the edge of the graph provided there is enough room. Works for both log-scale and regular settings.}

\item{expLabels}{logical; prints the major tick labels is \eqn{logScale^{x}}{logScale^x} instead of the raw value}

\item{lWidth}{numeric; Line width, equivelent to the cex \code{lwd} option.}

\item{na.rm}{logical; Should \code{NA}s be removed from the data set? Both data input and the factor input from \code{by} with be checked.}

\item{verbose}{logical; Prints summary and p-value calculations to the screen. All data is silently by the function returned either way.}

\item{logAdjustment}{= numeric; This number is added to the input data prior to log transformation. Default value is 1.}

\item{xLim}{numeric vector; manually set the limits of the plotting area (eg. \code{xLim=c(min,max)}). Used to format the x-axis. by default but will modify the x-axis if \code{side=\link{TRUE}}.}

\item{yLim}{numeric vector; manually set the limits of the plotting area (eg. \code{yLim=c(min,max)}). Used to format the y-axis.}

\item{strictLimits}{logical; eliminates padding on the value axis so 0 can be flush with the x-axis. Defaults to \code{\link{FALSE}}.}

\item{legend}{logical/character; Draw a legend in the plot margins. If a character string is given it will overide the factor name default for the legend title.}

\item{trimCurves}{logical; Should the density curves be restricted to the min and max of the data.}

\item{...}{additional options for S3 method variants}
}
\description{
Draws a kernal density plot of one or two variables.
}
\details{
TODO
}
\examples{
TODO<-1
}
\seealso{
\code{\link[stats]{density}}, \code{\link[graphics]{contour}}, \code{\link[graphics]{persp}}, \code{\link[rgl]{persp3d}}, \code{\link[KernSmooth]{bkde}}
}
