% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\name{npColorTheme}
\alias{npColorTheme}
\title{NicePlots Theme: Colors}
\usage{
npColorTheme
}
\description{
A more colorful and sparse option compared to the default
}
\details{
This theme uses fill, line, and point colors. Not a good option for \code{pointHighlights}
}
