% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\name{npDefaultTheme}
\alias{npDefaultTheme}
\title{NicePlots Theme: Default}
\usage{
npDefaultTheme
}
\description{
This is the default theme for NicePlots
}
\details{
This default theme has uses transparent solid circles over matching backgrounds for point overlays with up to 8 colors.
}
