% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\docType{data}
\name{npStataTheme}
\alias{npStataTheme}
\title{Nice Plots Theme: Stata Like}
\format{An object of class \code{npTheme} (inherits from \code{list}) of length 46.}
\usage{
npStataTheme
}
\description{
Looks a bit like stata plots
}
\details{
Todo
}
\keyword{datasets}
