% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\name{npThemes}
\alias{npThemes}
\title{Find Themes}
\usage{
npThemes()
}
\description{
Find NicePlot Themes Availble in the Namespace
}
\details{
This function takes the output of \code{\link{search}()} to check for variables that are members of \code{npTheme} \link{class}.
These variable names are then returned by the function.
}
\examples{
npThemes()
newBasic<-basicTheme
newBasic$plotColors$lines=rainbow(10)
npThemes()
}
\seealso{
\code{\link{search}}, \code{\link{class}}
}
