% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_data_processing.R
\name{prepBarData}
\alias{prepBarData}
\title{Prepare and print basic statistics for niceBar}
\usage{
prepBarData(
  x,
  by,
  errorMultiple = 1,
  upperErrorFun = "sd",
  lowerErrorFun = upperErrorFun,
  aggFunction = "mean",
  stack = FALSE,
  subgroup = FALSE
)
}
\arguments{
\item{x}{list; a list object returned by \code{\link{prepCategoryWindow}}}

\item{by}{factor or dataframe of factors; One or more factors that control how the data is grouped. The first column is the primary grouping factor and the second and thrid columns are used for sub-grouping and/or stacking as needed.}

\item{errorMultiple}{numeric; How many standard errors/deviations should be represented by the error bars.}

\item{upperErrorFun}{character string; Determines how the error barse are calculated. Options are \code{\link[stats]{sd}} (standard deviation), \code{\link{se}} (standard error), \code{\link[base]{min}}, \code{\link[base]{max}} and \code{\link{boot95ci}} (bootstrap 95\% confidence interval).}

\item{lowerErrorFun}{character; A character vector for the primary group names}

\item{aggFunction}{character string; An string naming the function to be used to aggregate the grouped data. Typically should be either \code{\link[stats]{median}} or \code{\link[base]{mean}}.}

\item{stack}{logical; Should one of the factors in \code{by} be used make a stacked bar plot. Note that this sort of analysis is nonsensical for many data sets.}

\item{subgroup}{logical; If \code{\link{TRUE}} the data will be faceted into subgroups within the primary factor levels. Ignored if \code{by} is a \code{\link[base]{factor}} or \code{x} is a \code{\link[base]{data.frame}}.}
}
\description{
Processes the input data, factors, and options to produce summary data for drawing bar plots
}
\details{
This functions takes the numeric data from \code{x} and the factor information from code{by} along with all
of the plotting options to calculated aggregate measures (mean/median) and error bars associated with those
measures. The error options include standard deviation (\code{sd}), standard error of the meand (\code{se}),
\code{range}, 95\% confidence interval based on the t-statistic (\code{t95ci}) and empirical 95\% confidence
intervals of the median or mean (\code{boot95ci}). This pre-processing is used to determine the default axis
limits, the summary data included in the output and passed along to the plotting functions.
}
\examples{
data(iris)
#\donttest{myData<-prepBarData(iris[,1:4],by=iris$Species)}
#\donttest{print(myData)}

}
\seealso{
\code{\link{niceBar}}, \code{\link{boot95ci}}, \code{\link{drawBar}}
}
