% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_categorical_plot_setup.R
\name{prepCategoryWindow}
\alias{prepCategoryWindow}
\title{prepare a plotting environment for categorical data such as bar plots or box plots}
\usage{
prepCategoryWindow(
  x,
  by = NULL,
  groupNames = levels(by),
  minorTick = FALSE,
  guides = TRUE,
  yLim = NULL,
  rotateLabels = FALSE,
  rotateY = TRUE,
  theme = NA,
  plotColors = if (is.na(theme)) {     list(bg = "open", guides = "black", lines =
    "gray22", points = "darkgrey", fill = "white") } else {     theme$plotColors },
  trim = FALSE,
  logScale = FALSE,
  axisText = c(NULL, NULL),
  minorGuides = FALSE,
  extendTicks = F,
  subGroup = FALSE,
  expLabels = TRUE,
  sidePlot = FALSE,
  subGroupLabels = NULL,
  strictLimits = F,
  legend = FALSE,
  pointHighlights = FALSE,
  logAdjustment = 1,
  stack = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector or data frame; The input to \code{prepCategoryWindow} can be a numeric vector a  data frame of numeric vectors.}

\item{by}{factor or data frame of factors; used as the primary grouping factor and the factor levels will be used as group names if \code{groupNames} is not specified. If \code{by} is a data frame and \code{subGroup=\link{TRUE}}, the second column is assumed to be a secondary grouping factor, breaking out the data into sub-categories within each major group determined by the levels of the first column.}

\item{groupNames}{character vector; overrides the factor levels of \code{by} to label the groups}

\item{minorTick}{positive integer; number of minor tick-marks to draw between each pair of major ticks-marks.}

\item{guides}{logical; will draw guidelines at the major tick-marks if set to \code{\link{TRUE}}. Color of the guidelines is determined by \code{plotColors$guides}.}

\item{yLim}{numeric vector; manually set the limits of the plotting area (eg. \code{yLim=c(min,max)}). Used to format the y-axis by default but will modify the x-axis if \code{side=\link{TRUE}}.}

\item{rotateLabels}{logical; sets \code{las=2} for the x-axis category labels. Will affect y-axis if \code{side=\link{TRUE}}. Note that this may not work well if long names or with subgrouped data.}

\item{rotateY}{logical; sets \code{las=2} for the y-axis major tick-mark labels. Will affect x-axis if \code{side=\link{TRUE}}.}

\item{theme}{list; A \code{NicePlots} plotColor list from a theme.}

\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions. Names left unspecified will be added and set to default values automatically.}

\item{trim}{positive numeric; passed to \code{threshold} argument of \code{\link{quantileTrim}} if any data points are so extreme that they should be removed before plotting and downstream analysis. Set to \code{\link{FALSE}} to disable.}

\item{logScale}{positive numeric; the base for the for log scale data transformation calculated as \code{log(x+1,logScale)}.}

\item{axisText}{character; a length two character vector containing text to be prepended or appended to the major tick labels, respectively.}

\item{minorGuides}{logical; draws guidelines at minor tick-marks}

\item{extendTicks}{logical; extends minor tick-marks past the first and last major tick to the edge of the graph provided there is enough room. Works for both log-scale and regular settings.}

\item{subGroup}{logical; use additional column in \code{by} to group the data within each level of the major factor.}

\item{expLabels}{logical; prints the major tick labels is \eqn{logScale^{x}}{logScale^x} instead of the raw value}

\item{sidePlot}{logical; switches the axis to plot horizontally instead of vertically.}

\item{subGroupLabels}{character vector; sets the labels used for the \code{subGroup} factor. Defaults to the levels of the factor.}

\item{strictLimits}{logical; eliminates padding on the value axis so 0 can be flush with the x-axis. Defaults to \code{\link{FALSE}}.}

\item{legend}{logical/character; Draw a legend in the plot margins. If a character string is given it will overide the factor name default for the legend title.}

\item{pointHighlights}{logical; Is pointHightlights turned on? This is used to determin with column of \code{by} should be used for legend factor levels.}

\item{logAdjustment}{= numeric; This number is added to the input data prior to log transformation. Default value is 1.}

\item{stack}{logical; Used for stack stacked bar plots. Used exclusively by \code{\link{niceBar}}.}

\item{...}{additional options mostly to be passed along to subsequent functions}
}
\value{
formats the plotting area and returns a named list with 'data' and 'labels' corresponding to the trimmed and/or transformed data and the labels for the primary factors, respectively.
}
\description{
takes untransformed data and draws the x and y axis with support for subgrouping data within factors, log transformation and outlier trimming.
}
\details{
This function does all the hard work of setting up the x and y axis for plotting as well as optionally log transforming and/or trimming the data of outliers. In particular, it adds much more robust support for plotting of log transformed data and subgrouping of primary vectors. Other features include the addition of both major and minor guidelines, support for horizontal plotting and improved label formatting options.
}
\examples{
todo<-1

}
\seealso{
\code{\link[grDevices]{axisTicks}}, \code{\link[graphics]{axis}}, \code{\link{makeLogTicks}}, \code{\link{facetSpacing}}
}
