% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_categorical_plot_setup.R
\name{prepLegendMarigins}
\alias{prepLegendMarigins}
\title{Set Legend Margins}
\usage{
prepLegendMarigins(
  x,
  by,
  theme,
  legend,
  pointHighlights = FALSE,
  subGroup = TRUE,
  stack = FALSE,
  is2D = FALSE
)
}
\arguments{
\item{x}{numeric vector or data frame; The input to \code{prepCategoryWindow} can be a numeric vector a  data frame of numeric vectors.}

\item{by}{factor or data frame of factors; used as the primary grouping factor and the factor levels will be used as group names if \code{groupNames} is not specified. If \code{by} is a data frame and \code{subGroup=\link{TRUE}}, the second column is assumed to be a secondary grouping factor, breaking out the data into sub-categories within each major group determined by the levels of the first column.}

\item{theme}{list object; Themes are are an optional way of storing graphical preset options that are compatible with all nicePlot graphing functions.}

\item{legend}{character; Title for the legend collumn. Set to \code{\link{TRUE}} if no header is desired.}

\item{pointHighlights}{logical; Is pointHightlights turned on? This is used to determin with column of \code{by} should be used for legend factor levels.}

\item{subGroup}{subGroup logical; use additional column in \code{by} to group the data within each level of the major factor.}

\item{stack}{logical; Used for stack stacked bar plots. Used exclusively by \code{\link{niceBar}}.}

\item{is2D}{logical; Is this for a 2D scatterplot or density plot? The first column of \code{by} will be use if set to \code{\link{TRUE}}.}
}
\value{
Does not return a value but changes the global \code{par(mai)} settings.
}
\description{
Calculate margin needed to plot a NicePlots legend
}
\details{
This is a utility function used by the window setup routines to calculate how much space the legend will take up in the margin and setting the \code{par(mai)} setting accordingly.
}
\examples{
TODO<-1
}
