% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_data_processing.R
\name{prepNiceData}
\alias{prepNiceData}
\title{Prepare and print basic statistics for niceBox and niceVio}
\usage{
prepNiceData(prepedData, by, subGroup = FALSE, outliers = TRUE, filter,
  groupNames, plotLoc, width = 1, flipFacts = FALSE, verbose = FALSE)
}
\arguments{
\item{prepedData}{list; a list object returned by \code{\link{prepCategoryWindow}}}

\item{by}{factor or dataframe of factors; One or more factors that control how the data is grouped. The first column is the primary grouping factor and the second and thrid columns are used for sub-grouping and highlighting as needed.}

\item{subGroup}{logical; Should the data be faceted into subgroups within the primary factor levels. Ignored if \code{by} is a \code{\link[base]{factor}}.}

\item{outliers}{positive numeric; number of interquartile ranges (IQR) past the Q1 (25\%) and Q3 (75\%) cumulative distribution values. Outliers are often defined as \eqn{1.5 \times IQR}{1.5 * IQR} and extreme outliers are more than \eqn{3 \times IQR}{3 * IQR} away from the inner 50\% data range.}

\item{filter}{logical vector; Used to further filter the data if necissary.}

\item{groupNames}{character; A character vector for the primary group names}

\item{plotLoc}{numeric vector; A vector indicating where each element should be plotted}

\item{width}{numeric; A multiplier that controls how wide the ploting elements will be. Setting \code{width=1.1} would result in plot elements being 10\% wider.}

\item{flipFacts}{logical; When a dataframe of values is given, column names are used as a secondary grouping factor by default. Setting \code{flipFacts=\link{TRUE}} makes the column names the primary factor and \code{by} the secondary factor.}

\item{verbose}{logical; Will print summary statistics to the screen if set to \code{\link{TRUE}}. The function will return the calculations either way.}
}
\description{
Uses filtred data with subgroup and factor information to calculate quartile data for display and plotting.
}
\details{
To aid in data interpretation and exploration, quartile distribution statistics are calculated for each group and subgroup
if specified. For \code{\link{niceBox}} this data is also used to plot the data. The data is parsed by checking \code{outlier} and \code{subGroup} status
as weel as checking if either \code{prepedData} or \code{by} are a \code{\link[base]{data.frame}} or a \code{\link[base]{vector}}.
}
\examples{
data(iris)
filter<-rep(TRUE,length(iris$Species))
loc<-seq(1,length(levels(iris$Species)))
data<-list(data=iris[,1:4])
\donttest{myData<-prepNiceData(data,by=iris$Species,filter=filter,plotLoc=loc,
    groupNames=levels(iris$Species),outliers=FALSE)}
\donttest{print(myData)}

}
\seealso{
\code{\link{niceBox}}, \code{\link{niceVio}}, \code{\link{niceDots}}
}
