% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_categorical_plot_setup.R
\name{prepNiceWindow}
\alias{prepNiceWindow}
\title{prepare a plotting environment for non-categorical data such as scatter plots}
\usage{
prepNiceWindow(
  x,
  by = NULL,
  minorTick = FALSE,
  guides = TRUE,
  yLim = NULL,
  xLim = NULL,
  rotateLabels = FALSE,
  theme = NA,
  plotColors = if (is.na(theme)) {     list(bg = "open", guides = "black", lines =
    "gray22", points = "darkgrey", fill = "white") } else {     theme$plotColors },
  logScaleX = FALSE,
  logScaleY = FALSE,
  axisText = list(x = c(NULL, NULL), y = c(NULL, NULL)),
  minorGuides = FALSE,
  extendTicks = F,
  subGroup = FALSE,
  expLabels = TRUE,
  strictLimits = F,
  legend = FALSE,
  logAdjustment = 1,
  makePlot = TRUE
)
}
\arguments{
\item{x}{numeric vector or data frame; The input to \code{prepCategoryWindow} can be a numeric vector a  data frame of numeric vectors.}

\item{by}{factor or data frame of factors; used as the primary grouping factor and the factor levels will be used as group names if \code{groupNames} is not specified. If \code{by} is a data frame and \code{subGroup=\link{TRUE}}, the second column is assumed to be a secondary grouping factor, breaking out the data into sub-categories within each major group determined by the levels of the first column.}

\item{minorTick}{positive integer; number of minor tick-marks to draw between each pair of major ticks-marks.}

\item{guides}{logical; will draw guidelines at the major tick-marks if set to \code{\link{TRUE}}. Color of the guidelines is determined by \code{plotColors$guides}.}

\item{yLim}{numeric vector; manually set the limits of the plotting area (eg. \code{yLim=c(min,max)}). Used to format the y-axis.}

\item{xLim}{numeric vector; manually set the limits of the plotting area (eg. \code{xLim=c(min,max)}). Used to format the x-axis. by default but will modify the x-axis if \code{side=\link{TRUE}}.}

\item{rotateLabels}{logical; sets \code{las=2} for the x-axis category labels. Will affect y-axis if \code{side=\link{TRUE}}. Note that this may not work well if long names or with subgrouped data.}

\item{theme}{list; A \code{NicePlots} plotColor list from a theme.}

\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions. Names left unspecified will be added and set to default values automatically.}

\item{logScaleX}{positive numeric; the base for the for log scale data transformation for the x-axis calculated after the \code{logAdjustment} factor is added to the values.}

\item{logScaleY}{positive numeric; the base for the for log scale data transformation for the y-axis calculated after the \code{logAdjustment} factor is added to the values.}

\item{axisText}{list; A named list, "x" and "y" with two optional character values to be placed before and/or after the numeric axis label, respectively.}

\item{minorGuides}{logical; draws guidelines at minor tick-marks}

\item{extendTicks}{logical; extends minor tick-marks past the first and last major tick to the edge of the graph provided there is enough room. Works for both log-scale and regular settings.}

\item{subGroup}{logical; use additional column in \code{by} to group the data within each level of the major factor.}

\item{expLabels}{logical; prints the major tick labels is \eqn{logScale^{x}}{logScale^x} instead of the raw value}

\item{strictLimits}{logical; eliminates padding on the value axis so 0 can be flush with the x-axis. Defaults to \code{\link{FALSE}}.}

\item{legend}{logical/character; Draw a legend in the plot margins. If a character string is given it will overide the factor name default for the legend title.}

\item{logAdjustment}{numeric; This number is added to the input data prior to log transformation. Default value is 1.}

\item{makePlot}{logical; This format the data and ploting area without drawing anything if set to \code{\link{FALSE}}.}
}
\value{
formats the plotting area and returns a named list with 'data' and 'labels' corresponding to the trimmed and/or transformed data and the labels for the primary factors, respectively.
}
\description{
takes untransformed data and draws the x and y axis with support for log transformation and outlier trimming.
}
\details{
This function does all the hard work of setting up the x and y axis for plotting as well as optionally log transforming and/or trimming the data of outliers. In particular, it adds much more robust support for plotting of log transformed data. Other features include the addition of both major and minor guidelines, support for horizontal plotting and improved label formatting options.
}
\examples{
todo<-1

}
\seealso{
\code{\link[grDevices]{axisTicks}}, \code{\link[graphics]{axis}}, \code{\link{makeLogTicks}}, \code{\link{facetSpacing}}
}
