% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicePlots2.R
\name{procNiceOptions}
\alias{procNiceOptions}
\title{Process ploting options}
\usage{
procNiceOptions(x, by, minorTick, pointShape, wiskerLineType, lWidth, capWidth,
  pointLaneWidth, width, guides, pointSize, subGroup = FALSE, stack = F,
  pointHighlights = F, type = c("BP", "VP", "DP", "Bar"), theme, plotColors,
  pointMethod, logScale, drawPoints, groupNames, swarmOverflow,
  errorCap = NULL)
}
\arguments{
\item{x}{Data to be plotted which has been preprocessed by \code{\link{dataFlightCheck}}}

\item{by}{factor or dataframe of factors; One or more factors that control how the data is grouped. The first column is the primary grouping factor and the second and thrid columns are used for sub-grouping and highlighting as needed.}

\item{minorTick}{numeric; Number of minor tickmarks to be drawn between the major marks}

\item{pointShape}{numeric; vector of numbers corresponding to pty options for ploting data overlays.}

\item{wiskerLineType}{numeric; number corresponding to lty option for drawing the whiskers and error bars for box plots and bar plots, respectively.}

\item{lWidth}{numeric; number creesponding to the lwd option for ploting lines on the graph}

\item{pointLaneWidth}{numeric; This controls how far data point dots can move along the categorical axis when plotting. Used for \code{pointMethod} options 'jitter', 'beeswarm', and 'distribution'.}

\item{width}{numeric; A multiplier that controls how wide the ploting elements will be. Setting \code{width=1.1} would result in plot elements being 10\% wider.}

\item{guides}{logical; Should guidelines be drawn at the major tick marks.}

\item{pointSize}{numeric; vector of numerics controling the size of points on the data overlay}

\item{subGroup}{logical; Should the data be faceted into subgroups within the primary factor levels. Ignored if \code{by} is a \code{\link[base]{factor}}.}

\item{stack}{logical; Triggers stacked bar analysis for bar plots}

\item{pointHighlights}{logical; will use additional factors in \code{by} to highlight points in the dot plot}

\item{type}{character; What kind of plot is this for? Case sensitive options are "BP", "DP", "VP", and "Bar" corresponding to box plots, dot plots, violin plots, and bar plots, respectively.}

\item{theme}{list object; Themes are are an optional way of storing graphical preset options that are compatible with all nicePlot graphing functions.}

\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions. Names left unspecified will be added and set to default values automatically.}

\item{pointMethod}{character; method to be used for ploting dots. Can be set to "jitter", "linear", "beeswarm" or "distribution".}

\item{logScale}{numeric; Should a log scale use used (\code{TRUE}/\code{FALSE})? Otherwise indicates the base for the log transformation.}

\item{drawPoints}{logical; draws a dot plot overlay of the data.}

\item{groupNames}{character; A character vector for the primary group names}

\item{swarmOverflow}{character; Valid options are: "none", "wrap", "gutter", "random", and "omit". Controls how to wantly point stacks that would overflow the pointLaneWidth option.}

\item{errorCap}{character; Determines the style for the ends of the error bars. Valid options are \code{ball}, \code{bar} or \code{none}.}

\item{capWidth;}{numeric; Width of the cap relative to the bar/box width for box plots and bar plots.}
}
\value{
Named listed of graphical options
}
\description{
Integrates theme and user arguments to finalize all options prior to plotting
}
\details{
This is a private utility function used by NicePlots to integrate user options with theme defaults.
Anything specified by the user is treated literally with no additional optimization.
Colors, point shapes, fills, etc. are optimized to best enhance the relevant factor visualisations selected.
The finalized parameters are returned as a named list.
}
\seealso{
\code{\link{formatPlotColors}}, \code{\link{niceBox}}, \code{\link{niceDots}}, \code{link{niceVio}}, \code{\link{niceBar}}
}
