% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_utility.R
\name{se}
\alias{se}
\title{Calculate the standard error of the mean}
\usage{
se(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a double corresponding to the standard error of the mean
}
\description{
\code{se} takes a numeric vector and returns the corresponding standard error of the mean
}
\details{
This is a convenience function internal to \code{NicePlots} and is not exported. If \code{x} is a numeric vector, \code{SD} is the standard
deviation of \code{x} and  \code{N} is the length of \code{x} then the standard error (se) of the mean may be caclulated as:
\deqn{se = \frac{SD}{\sqrt{N}}}
}
\examples{
data(iris)
#se(iris$Sepal.Length)
}
\seealso{
\code{\link[stats]{sd}}
}
