% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_utility.R
\name{setAlpha}
\alias{setAlpha}
\title{add alpha transparency to a named color}
\usage{
setAlpha(x, alpha = 0.2)
}
\arguments{
\item{x}{character string; a text string corresponding to an R color}

\item{alpha}{numeric [0-1]; sets the level of transparency.}
}
\value{
An rbg color with transparancy \code{alpha}.
}
\description{
Takes a named color such as "red" or "darkgreen" and adds a level of transparency based on the alpha setting.
}
\details{
\code{setAlpha} is a convenience function that uses the \code{\link[grDevices]{col2rgb}} and \code{\link[grDevices]{rgb}} to add transparancy to named colors.
}
\examples{
plot(1,1,col="white",xlim=c(0,10),ylim=c(0,10))
rect(1,1,7,7,col=setAlpha("darkblue"))
rect(3,3,9,9, col=setAlpha("red"))

}
\seealso{
\code{\link{makeColorMatrix}}, \code{\link[grDevices]{rainbow}}, \code{\link[grDevices]{col2rgb}}, \code{\link[grDevices]{rgb}}.
}
