% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_utility.R
\name{t95ci}
\alias{t95ci}
\title{Calculate a 95\% confidence interval from the t-distribution}
\usage{
t95ci(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a double corresponding to the length of one arm ofthe 95\% confidence interval
}
\description{
\code{se} Calculates the 95\% confidence interval of the mean for a vector of data based on the t-distribution
}
\details{
This is a convenience function internal to \code{NicePlots} and is not exported. If \code{x} is a numeric vector, \code{SD} is the standard
deviation of \code{x}, \code{N} is the length of \code{x} and \code{TQ} is the t-distrubition quantile for \code{0.975} with \code{N-1}
degrees for freedom  then the 95\% confidence interval of the mean may be caclulated as:
\deqn{95\%ci = \frac{QT*SD}{\sqrt{N}}}
}
\examples{
data(iris)
#t95ci(iris$Sepal.Length)
}
\seealso{
\code{\link[stats]{sd}}
}
