% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\docType{data}
\name{basicTheme}
\alias{basicTheme}
\title{NicePlots Theme: Basic}
\format{An object of class \code{npTheme} (inherits from \code{list}) of length 54.}
\usage{
basicTheme
}
\description{
This is the default theme for nicePlots
}
\details{
This default theme has uses transparent solid circles for point overlays with up to 8 colors.
Fill and line options as constant.
}
\keyword{datasets}
