% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_categorical_plot_setup.R
\name{makeNiceLegend}
\alias{makeNiceLegend}
\title{Draw a nice plot legened}
\usage{
makeNiceLegend(labels, title = "Legend", fontCol = "black",
  border = NULL, lineCol = NA, bg = NA, col = makeColorMatrix()[,
  3], shape = "rect", size = 0.66, spacing = 0.2)
}
\arguments{
\item{labels}{character vector; The names of the levels decribed in the legend. Typically factor levels.}

\item{title}{character; The title of the legend. This defaults to "Legend" if unspecificed.}

\item{fontCol}{R color; Color of the legend text.}

\item{border}{R color; The color of the rectanglar border surrounding the legend. Defaults to \code{\link{NULL}} which supresses this feature}

\item{lineCol}{R color; The color of the line colors for the color key. Optional. Defaults to \code{\link{NA}}.}

\item{bg}{R color; Sets the background color for the legend aread. Note that this can be distinct the the margin background.}

\item{col}{R color vector; A vector of colors determining the color of the color code boxes.}

\item{shape}{character; Determins if the color code is rectangles or circles. Valid ptions are "rect", "rectangle", "circ", or "circle". Not there is no funcitonal difference between the synonyms.}

\item{size}{numeric; Sets the legend font cex sizing.}

\item{spacing}{numeric; Determins the total amount of padding (sum of upper and lower padding) surrounding each line. in the legend in units of font line hight.}
}
\description{
Draws a customizable legend in the margins based on factor levels.
}
\details{
This functions works with plot enviroment initializing functions such as \code{\link{prepCategoryWindow}}
to expand the right margin to accomodate a figure legend.
}
\examples{
ToDo<-1

}
\seealso{
\code{\link{legend}}, \code{\link{prepCategoryWindow}}, \code{\link{niceBox}}, \code{\link{niceDots}}, \code{\link{niceBar}}, \code{\link{niceVio}}
}
