% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_niceDensity.R
\name{niceDensity}
\alias{niceDensity}
\title{draw a kernal density plot}
\usage{
niceDensity(
  x,
  by = NULL,
  drawPoints = TRUE,
  groupNames = NULL,
  subgroup = FALSE,
  bandwidth = NULL,
  drawRug = FALSE,
  useRgl = FALSE,
  plotType = c("contour", "surface"),
  theme = basicTheme,
  main = NULL,
  sub = NULL,
  ylab = NULL,
  xlab = NULL,
  minorTick = FALSE,
  guides = NULL,
  plotColors = NULL,
  logScale = FALSE,
  axisText = c(NULL, NULL),
  rotateLabels = FALSE,
  add = FALSE,
  minorGuides = NULL,
  extendTicks = TRUE,
  expLabels = FALSE,
  lWidth = NULL,
  na.rm = FALSE,
  verbose = FALSE,
  logAdjustment = 1,
  xLim = NULL,
  yLim = NULL,
  strictLimits = FALSE,
  legend = FALSE,
  trimCurves = TRUE,
  sidePlot = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector or data frame; The input to \code{prepCategoryWindow} can be a numeric vector a  data frame of numeric vectors.}

\item{by}{factor or data frame of factors; used as the primary grouping factor and the factor levels will be used as group names if \code{groupNames} is not specified. If \code{by} is a data frame and \code{subgroup=\link{TRUE}}, the second column is assumed to be a secondary grouping factor, breaking out the data into sub-categories within each major group determined by the levels of the first column.}

\item{drawPoints}{logical; draws a dot plot overlay for contour plots}

\item{groupNames}{character vector; overrides the factor levels of \code{by} to label the groups}

\item{subgroup}{logical; Will use the factor levels in \code{by} to plot a series of distributions from the data in \code{x}.}

\item{bandwidth}{numeric; Manually sets the bandwith for the kernal density estimation overiding default calculations. For 2D plots \code{bandwidth} should be a vector of length 2. Set to \code{\link{NULL}} or \code{\link{NA}} to enable default calculations.}

\item{drawRug}{logical; adds a \code{\link[graphics]{rug}} plot overlay to a kernel density plot. Default is FALSE.}

\item{useRgl}{logical; Should the library \code{\link[rgl]{rgl}} be used to make 3D surface plots.}

\item{plotType}{character; Can be set to \code{contour} or \code{surface} to control the type of 2D plot.}

\item{theme}{list object; Themes are are an optional way of storing graphical preset options that are compatible with all nicePlot graphing functions.}

\item{main}{character; title for the graph which is supplied to the \code{main} argument.}

\item{sub}{character; subtitle for the graph which is supplied to the \code{sub} argument.}

\item{ylab}{character; y-axis label.}

\item{xlab}{character; x-axis label.}

\item{minorTick}{positive integer; number of minor tick-marks to draw between each pair of major ticks-marks.}

\item{guides}{logical; will draw guidelines at the major tick-marks if set to \code{\link{TRUE}}. Color of the guidelines is determined by \code{plotColors$guides}.}

\item{plotColors}{list; a named list of vectors of colors that set the color options for all NicePlot functions. Names left unspecified will be added and set to default values automatically.}

\item{logScale}{numeric; A length two numeric vector indicating the log scale the x and y axis, respectively. If only one value is given, it is applied to boxth axis. Set to \code{\link{FALSE}} to disable (Default).}

\item{axisText}{list; A named list, "x" and "y" with two optional character values to be placed before and/or after the numeric axis label, respectively.}

\item{rotateLabels}{logical; sets \code{las=2} for the x-axis category labels. Will affect y-axis if \code{side=\link{TRUE}}. Note that this may not work well if long names or with subgrouped data.}

\item{add}{logical; Adds the plot to the existing window rather the generating a new plotting enviroment}

\item{minorGuides}{logical; draws guidelines at minor tick-marks}

\item{extendTicks}{logical; extends minor tick-marks past the first and last major tick to the edge of the graph provided there is enough room. Works for both log-scale and regular settings.}

\item{expLabels}{logical; prints the major tick labels is \eqn{logScale^{x}}{logScale^x} instead of the raw value}

\item{lWidth}{numeric; Line width, equivelent to the cex \code{lwd} option.}

\item{na.rm}{logical; Should \code{NA}s be removed from the data set? Both data input and the factor input from \code{by} with be checked.}

\item{verbose}{logical; Prints summary and p-value calculations to the screen. All data is silently by the function returned either way.}

\item{logAdjustment}{numeric; This number is added to the input data prior to log transformation. Default value is 1.}

\item{xLim}{numeric vector; lower a upper limits for the x-axis.}

\item{yLim}{numeric vector; manually set the limits of the plotting area (eg. \code{yLim=c(min,max)}). Used to format the y-axis.}

\item{strictLimits}{logical; eliminates padding on the value axis so 0 can be flush with the x-axis. Defaults to \code{\link{FALSE}}.}

\item{legend}{logical/character; Draw a legend in the plot margins. If a character string is given it will overide the factor name default for the legend title.}

\item{trimCurves}{logical; Should the density curves be restricted to the min and max of the data.}

\item{sidePlot}{logical; Swaps the x and y axis for ploting. Does not apply to 2D and 3D plots.}

\item{...}{additional options for S3 method variants}
}
\description{
Draws a kernal density plot of one or two variables.
}
\details{
\code{niceDensity} is draws gernal density plots. This can be a single kernal density plot or a series of density plots overplotted
on the graph based on a grouping factor from \code{by}. The \code{subgroup} option should be set to \code{\link{TRUE}} to anable this functionality.
If two or more variables are passed in \code{x} in a \code{\link{data.frame}}, \code{\link{matrix}} or \code{\link[tibble]{tibble}}, then the
first two columns will be used to generate a scatter plot with a 2-D scatter plot with a \code{\link[graphics]{contour}} plot overlay giving the 2D
kernal density estimates. Setting \code{subgroup} to \code{\link{TRUE}} is this case will color the points by the first factor in \code{by}.
Alternatively, for two column input,  \code{plotType} can be set to "suface" to draw a 3D prepresnetation of the kernal density over the x-y plane using \code{\link[graphics]{persp}}.
Plotting options such as \code{theta}/\code{phi} to control rotation for \code{\link[graphics]{persp}} or \code{nlevels} for \code{\link[graphics]{contour}}, can be entered as arguments to \code{niceDensity} and will be passed along via \code{\link{do.call}}.
Finally, if the \code{rgl} package is installed and \code{useRgl} is set to \code{\link{TRUE}}, the interactive \code{\link[rgl]{persp3d}} function will be used instead.
All kernal densities are calculated using the \code{KernSmooth} package. The values in \code{npData} object returned by the function are slightly different.
The "summary" value gives the min and max for \code{x} and optionally \code{y} along with the density estimate \code{fhat}.
The "stats" value contains a \code{\link{data.frame}} the actual \code{x}, \code{y}, and \code{fhat} paired values.
The theme value \code{curvePoints} is used to determine the number of line segments per curve and can be overriden in the command line like all other theme options.
}
\examples{
data(iris)
#Basic kernal density plot
niceDensity(iris[,1], main="Distrinbution of Sepal Length")

#Kernal density plot subgrouped by species
niceDensity(iris[,1],iris$Species, main="Sepal Length by Species",
subgroup=TRUE)

#Same thing with trimCurves=FALSE
niceDensity(iris[,1],iris$Species, main="Sepal Length by Species",
subgroup=TRUE, trimCurves=FALSE)

#2D Density of Sepal Width vs. Length
niceDensity(iris[,1:2],iris$Species, main="Sepal Width vs. Length",
subgroup=TRUE)

#Representing the 2D contour plot in 3D
niceDensity(iris[,1:2],iris$Species, main="Sepal Width vs. Length",
plotType="surface", theta=-60)

}
\seealso{
\code{\link[stats]{density}}, \code{\link[graphics]{contour}}, \code{\link[graphics]{persp}}, \code{\link[rgl]{persp3d}}, \code{\link[KernSmooth]{bkde}}
}
