% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\docType{data}
\name{npColorTheme}
\alias{npColorTheme}
\title{Nice Plots Theme: Colors}
\format{An object of class \code{npTheme} (inherits from \code{list}) of length 43.}
\usage{
npColorTheme
}
\description{
A more colorful and sparse option compared to the default
}
\details{
This theme uses fill, line, and point colors. Not a good option for \code{pointHighlights}
}
\keyword{datasets}
