% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\docType{data}
\name{npDefaultTheme}
\alias{npDefaultTheme}
\title{NicePlots Theme: Default}
\format{
An object of class \code{npTheme} (inherits from \code{list}) of length 51.
}
\usage{
npDefaultTheme
}
\description{
This is the default theme for NicePlots
}
\details{
This default theme has uses transparent solid circles over matching backgrounds for point overlays with up to 8 colors.
}
\keyword{datasets}
