% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceThemes.R
\docType{data}
\name{npGGTheme}
\alias{npGGTheme}
\title{NicePlots Theme: ggPlot Like}
\format{An object of class \code{npTheme} (inherits from \code{list}) of length 50.}
\usage{
npGGTheme
}
\description{
Looks a bit like stata plots
}
\details{
Todo
}
\keyword{datasets}
