% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_data_processing.R
\name{prepBarData}
\alias{prepBarData}
\title{Prepare and print basic statistics for niceBar}
\usage{
prepBarData(
  x,
  by,
  errorMultiple = 1,
  upperErrorFun = "sd",
  lowerErrorFun = upperErrorFun,
  aggFunction = "mean",
  stack = FALSE,
  subGroup = FALSE
)
}
\arguments{
\item{x}{list; a list object returned by \code{\link{prepCategoryWindow}}}

\item{by}{factor or dataframe of factors; One or more factors that control how the data is grouped. The first column is the primary grouping factor and the second and thrid columns are used for sub-grouping and/or stacking as needed.}

\item{errorMultiple}{numeric; How many standard errors/deviations should be represented by the error bars.}

\item{upperErrorFun}{character string; Determines how the error barse are calculated. Options are \code{\link[stats]{sd}} (standard deviation), \code{\link{se}} (standard error), \code{\link[base]{min}}, \code{\link[base]{max}} and \code{\link{boot95ci}} (bootstrap 95\% confidence interval).}

\item{lowerErrorFun}{character; A character vector for the primary group names}

\item{aggFunction}{character string; An string naming the function to be used to aggregate the grouped data. Typically should be either \code{\link[stats]{median}} or \code{\link[base]{mean}}.}

\item{stack}{logical; Should one of the factors in \code{by} be used make a stacked bar plot. Note that this sort of analysis is nonsensical for many data sets.}

\item{subGroup}{logical; If \code{\link{TRUE}} the data will be faceted into subgroups within the primary factor levels. Ignored if \code{by} is a \code{\link[base]{factor}} or \code{x} is a \code{\link[base]{data.frame}}.}
}
\description{
Processess the input data, factors, and options to produce summary data for drawing bar plots
}
\details{
todo<-1
}
\examples{
todo<-1

}
\seealso{
\code{\link{niceBar}}, \code{\link{boot95ci}}, \code{\link{drawBar}}
}
