% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_categorical_plot_setup.R
\name{prepLegendMarigins}
\alias{prepLegendMarigins}
\title{Set Legend Margins}
\usage{
prepLegendMarigins(x, by, theme, legend, pointHighlights = FALSE,
  subGroup = TRUE)
}
\arguments{
\item{x}{numeric; a numeric vector with the min and max values for the data set prior to log transformation.}

\item{by}{positive integer; the number of minor tick marks to be drawn between each major tick.}

\item{theme}{numeric; the logarithm base to use for the log scale transformation.}

\item{legend}{character; Title for the legend collumn. Set to \code{\link{TRUE}} if no header is desired.}

\item{pointHighlights}{character; a length two character vector containing text to be prepend or append to the major tick labels, respectively.}

\item{subGroup}{logical; if set to \code{\link{TRUE}}, the major labels will written as \eqn{logbase^{x}}{logbase^x}. Otherwise the labels will correspond to the non-transformed values at that point.}
}
\value{
Does not return a value but changes the global \code{par(mai)} settings.
}
\description{
Calculate margin needed to plot a NicePlots legend
}
\details{
This is a utility function used by the window setup routines to calculate how much space the legend will take up in the margin and setting the \code{par(mai)} setting accordingly.
}
\examples{
TODO<-1
}
